Node Visitor for spreads
This checks spreads and converts to an array or hash spread if it's able
to determine it at compile time. This can't be done with variables and
still needs to go through the argument spreader, but this should improve
performance a lot by not needing to go through that at runtime.